[Ide] Move native splitter dragging code to an abstract class
