+++ main/src/core/MonoDevelop.Ide/MonoDevelop.Ide.csproj
@@ -4250,6 +4250,7 @@
     <Compile Include="MonoDevelop.Components.Docking\MacSplitterWidget.cs" />
     <Compile Include="MonoDevelop.Components.Docking\SplitterWidgetWrapper.cs" />
     <Compile Include="MonoDevelop.Components.Docking\SplitterMacHostWidget.cs" />
+    <Compile Include="MonoDevelop.Components\Mac\DragEventTrapView.cs" />
   </ItemGroup>
   <ItemGroup Condition="'$(Configuration)' == 'DebugMac' OR '$(Configuration)' == 'ReleaseMac'">
     <Compile Include="MonoDevelop.Components\Mac\KeyCodes.cs" />
